Add Book table inside the shchema file, and create a new file for BookStorage module
