+++ main.rb
@@ -3,12 +3,12 @@ require_relative 'add_options'
 require_relative 'music_album'
 require_relative 'genre'
 require_relative 'author'
-require_relative 'book'
-require 'json'
+require_relative 'book_storage'
 
 class App
   include ListOptions
   include AddOptions
+  include BookStorage
 
   INPT_MSG = 'Enter your option number here --> '.freeze
 
@@ -68,23 +68,6 @@ class App
     @authors.each do |author|
       puts "Id: #{author.id} | First name: #{author.first_name}| Last name: #{author.last_name}"
     end
-  end
-
-  def parse_books
-    File.open('books.json', 'w') { |f| f.write JSON.generate([]) } unless File.exist? 'books.json'
-
-    JSON.parse(File.read('books.json')).map do |book|
-      @books << Book.new(book['publisher'], book['cover_state'], book['publish_date'])
-    end
-  end
-
-  def save_books
-    @json_books = []
-    @books.each do |book|
-      @json_books.push({ 'publisher' => book.publisher, 'cover_state' => book.cover_state,
-                         'publish_date' => book.publish_date })
-    end
-    File.write('books.json', JSON.generate(@json_books))
   end
 
   def exit_app
